Implement left pivot search and optimize processing
Added `pivots_search_left` function for efficient left pivot searching in sparse matrices, which does not need the full transpose. In the RREF, if the number of new pivots is too less, it is very expansive to compute the transpose of the matrix, so we only search the left columns.